Create specif rule for input objects to validate Lists of non-null
